moved tests into smoke-tests to prevent duplicating createTestTargets between multiple smoke test files
